requirements.txt

fastapi==0.115.8
uvicorn==0.34.0
langchain==0.3.19
langchain_groq==0.2.4
google-generativeai==0.8.4

langchain-community==0.3.17
unstructured==0.16.21
pymupdf==1.25.3  

google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
google-api-python-client==2.161.0
google-cloud-storage==2.10.0

python-docx==1.1.2
xhtml2pdf==0.2.16
pypandoc==1.15

duckduckgo-search==7.4.2
arxiv==2.1.3

python-dotenv==1.0.1
pydantic==2.10.6
requests==2.32.3

httpx==0.28.1
sqlalchemy==2.0.38
passlib==1.7.4
jose==1.0.0
pyjwt==2.10.1
python-jose==3.4.0
python-multipart==0.0.20
bcrypt==4.3.0
